fix(docsearch): make patch work
