+++ pages/_document.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @next/next/no-css-tags, @next/next/no-sync-scripts */
 import { Head, Html, Main, NextScript } from "next/document";
 
-import "../lib/docsearch";
+const ALGOLIA_HOST = `https://${process.env.NEXT_PUBLIC_ALGOLIA_APP_ID ?? "JAJUDFJBI4"}-dsn.algolia.net`;
 
 export default function Document() {
   return (
@@ -11,11 +11,9 @@ export default function Document() {
           <script src={process.env.NEXT_PUBLIC_REACT_DEVTOOLS_ADDRESS} />
         ) : null}
 
-        <link
-          rel="preconnect"
-          href={`https://${process.env.NEXT_PUBLIC_ALGOLIA_APP_ID ?? "JAJUDFJBI4"}-dsn.algolia.net`}
-          crossOrigin="anonymous"
-        />
+        <link rel="preconnect" href={ALGOLIA_HOST} crossOrigin="anonymous" />
+
+        <script src="/docsearch.js" data-algolia-host={ALGOLIA_HOST}></script>
       </Head>
       <body>
         <Main />
